+++ src/ddzy/utility/index.ts
@@ -1,12 +1,14 @@
 import utilityDOM from './dom/index';
 import utilityFunction from './function/index';
 import utilityNumber from './number/index';
+import utilityOthers from './others/index';
 
 
 const utility = {
   utilityDOM,
   utilityFunction,
   utilityNumber,
+  utilityOthers,
 };
 
 export default utility;